requirements.txt
asgiref==3.5.2
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
Django==4.1.3
mysqlclient==2.1.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
django-debug-toolbar==3.8.1
djoser==2.1.0
pillow==9.3.0
django-jazzmin==2.6.0
django-filter==22.1
django-cors-headers==3.13.0
celery==5.2.7
whitenoise==6.2.0
dj-database-url==1.2.0
django-redis==5.2.0
django-request-logging==0.7.5
gunicorn==20.1.0
django-storages==1.13.2
boto3==1.26.37